Added Analysis-72 4.0
The first release since 2018, Analysis-72 4.0 is a basic recoding of Analysis-72 3.1 which contains under-the-hood improvements, a new (original) icon, and proper use of multiple file extensions, with the Analysis-72 2.0 file format (*.data_analyze) being removed due to compatibility with older systems. Plus, Analysis-72 is now 32-bit, allowing for better compatibility.